fix: do not delete the state if it is the desired state
fixes #219

+++ lib/plugins/notifiers/open_close_threads/open_close_threads.rb
@@ -109,7 +109,7 @@ class OpenCloseThreadPlugin < Plugin
             ", '" + state + "')"
           )
         end
-      else
+      elsif rslt.first['state'] != state
         CfThread.connection.execute(
           "DELETE FROM opened_closed_threads WHERE user_id = " +
           current_user.user_id.to_s +
